feat: changed all Float32 to Float64, random seed

+++ src/homological_programming/SheafNodes.jl
@@ -9,37 +9,37 @@ abstract type AbstractSheafNode end
 
 #TODO: Refactor to make these not have explicit state
 mutable struct ThreadedSheafNode <: AbstractSheafNode
-    id::Int32
-    dimension::Int32
-    neighbors::Dict{Int32,AbstractMatrix}
-    in_channels::Dict{Int32,Channel}
-    out_channels::Dict{Int32,Channel}
-    x::Vector{Float32}
+    id::Int64
+    dimension::Int64
+    neighbors::Dict{Int64,AbstractMatrix}
+    in_channels::Dict{Int64,Channel}
+    out_channels::Dict{Int64,Channel}
+    x::Vector{Float64}
 end
 
 mutable struct AsyncSheafNode <: AbstractSheafNode
-    id::Int32
-    dimension::Int32
-    neighbors::Dict{Int32,AbstractMatrix}
-    in_channels::Dict{Int32,Channel}
-    out_channels::Dict{Int32,Channel}
-    x::Vector{Float32}
-    period::Int32
-    phase::Int32
-    iteration::Int32
-    # itersSinceUpdate::Int32
+    id::Int64
+    dimension::Int64
+    neighbors::Dict{Int64,AbstractMatrix}
+    in_channels::Dict{Int64,Channel}
+    out_channels::Dict{Int64,Channel}
+    x::Vector{Float64}
+    period::Int64
+    phase::Int64
+    iteration::Int64
+    # itersSinceUpdate::Int64
 end
 
 mutable struct DistributedSheafNode <: AbstractSheafNode
-    id::Int32
-    dimension::Int32
-    neighbors::Dict{Int32,AbstractMatrix}
-    in_channels::Dict{Int32,RemoteChannel}
-    out_channels::Dict{Int32,RemoteChannel}
-    x::Vector{Float32}
+    id::Int64
+    dimension::Int64
+    neighbors::Dict{Int64,AbstractMatrix}
+    in_channels::Dict{Int64,RemoteChannel}
+    out_channels::Dict{Int64,RemoteChannel}
+    x::Vector{Float64}
 end
 
-function add_neighbor!(s::AbstractSheafNode, n_id::Int32, restriction_map)
+function add_neighbor!(s::AbstractSheafNode, n_id::Int64, restriction_map)
     s.neighbors[n_id] = restriction_map
 end
